Working around the problem of unlined CUDA
